cloud_function/materialize-master-llm/requirements.txt
google-cloud-storage>=2.10.0
functions-framework>=3.5.0
